Q2_Video_Chat/requirements.txt
streamlit
moviepy
openai-whisper
sentence-transformers
faiss-cpu
openai
numpy
ffmpeg-python
requests
